// clients/pricingService/models/DiscountCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DiscountCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    description?: string | null;
    beginQuantity?: number;
    endQuantity?: number;
    percent?: number;
    value?: number;
    tenantId?: string | null;
    enrolmentId?: string | null;
    discountListId?: string | null;
};