// clients/pricingService/models/DiscountListCreateDto.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type DiscountListCreateDto = {
    readonly id?: string;
    readonly timestamp?: string;
    name?: string | null;
    currencyId?: string | null;
    tenantId?: string | null;
    enrolmentId?: string | null;
};